refactor : onboarding session 데이터 분리
